* Move AdvancedRecipe * Add the tape measure! * Add a StringUtil so that console space padding can work nicely. * Add a null check, remove debug * Fix console space padding * Add the tape measure * Add a Handler#itemStackEquals to make sure that item stacks are the same (Bukkit does this terribly so I am doing this myself) * Fix names of categories * Add Handler and make the getRecipe not default (For some reason it wasn't in EasierMCBase) * Change the EnderSword again, think this is a little better. * Make the AdvancedCraftingTable work!!!!!!!!!!!!! Whoooo!! * Made the ItemStackBuilder take an int, just easier than casting to short every time.

+++ src/main/java/com/bwfcwalshy/easiermc/utils/ItemStackBuilder.java
@@ -44,8 +44,8 @@ public class ItemStackBuilder {
         im = is.getItemMeta();
     }
 
-    public ItemStackBuilder setData(short data){
-        is.setDurability(data);
+    public ItemStackBuilder setData(int data){
+        is.setDurability((short) data);
         return this;
     }
